Fix docker build
The added `proc/` package was not added to the builder, causing the build
to fail. This change copies everything into the container, which is a
small overhead for not having to have this problem later again. Note that
nothing of the additionally copied files will show up in the final image
as the multi-stage build only copies the binary into the actual image.