Interpret AWS Access Denied errors as NOTFOUND
When running GET requests across multiple sources, it often happens that
items from one account are not available to read from another account.
That is not an error, but only indicates that the resource (item) is not
for the current account.

As downside this will hide errors where the IAM role/policy for the source
is misconfigured. Instead of `error processing: internal: all queries failed`,
now the user will just get an empty result.