Simplified AWS configuration
In this instance I have changed the way the AWS configuration gets set up to rely more on the automatic configuration that comes from the library rather than from environment variables that are set by the user. This means that the user can use any of the AWS configuration methods that are available to them, including a shared credentials file, environment variables, SSO Login, instance profile etc. and the tests will read from this automatically.