Capture a cleaner service version
This means that we get either the SHA of of the commit, or the tag, rather than the somewhat confusing `heads/main-0-gd4c5dc9` that we get at the moment. This is especially confusing since the SHA is the 7 digits, not 8 (i.e. `d4c5dc9` not `gd4c5dc9`)